+++ ci/scripts/check_ccache.sh
@@ -94,4 +94,3 @@ do
 done
 
 echo "could not download cache" && exit 1
-