+++ core/src/storage/IOReader.h
@@ -11,6 +11,7 @@
 
 #pragma once
 
+#include <memory>
 #include <string>
 
 namespace milvus {
@@ -18,9 +19,8 @@ namespace storage {
 
 class IOReader {
  public:
-    explicit IOReader(const std::string& name) : name_(name) {
-    }
-    ~IOReader() = default;
+    virtual void
+    open(const std::string& name) = 0;
 
     virtual void
     read(void* ptr, size_t size) = 0;
@@ -31,9 +31,11 @@ class IOReader {
     virtual size_t
     length() = 0;
 
- public:
-    std::string name_;
+    virtual void
+    close() = 0;
 };
+
+using IOReaderPtr = std::shared_ptr<IOReader>;
 
 }  // namespace storage
 }  // namespace milvus