+++ core/src/storage/disk/DiskIOWriter.cpp
@@ -9,33 +9,33 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "storage/file/FileIOReader.h"
+#include "storage/disk/DiskIOWriter.h"
 
 namespace milvus {
 namespace storage {
 
-FileIOReader::FileIOReader(const std::string& name) : IOReader(name) {
-    fs_ = std::fstream(name_, std::ios::in | std::ios::binary);
-}
-
-FileIOReader::~FileIOReader() {
-    fs_.close();
+void
+DiskIOWriter::open(const std::string& name) {
+    name_ = name;
+    len_ = 0;
+    fs_ = std::fstream(name_, std::ios::out | std::ios::binary);
 }
 
 void
-FileIOReader::read(void* ptr, size_t size) {
-    fs_.read(reinterpret_cast<char*>(ptr), size);
-}
-
-void
-FileIOReader::seekg(size_t pos) {
-    fs_.seekg(pos);
+DiskIOWriter::write(void* ptr, size_t size) {
+    fs_.write(reinterpret_cast<char*>(ptr), size);
+    len_ += size;
 }
 
 size_t
-FileIOReader::length() {
-    fs_.seekg(0, fs_.end);
-    return fs_.tellg();
+DiskIOWriter::length() {
+    return len_;
 }
+
+void
+DiskIOWriter::close() {
+    fs_.close();
+}
+
 }  // namespace storage
 }  // namespace milvus