+++ shards/mishards/__init__.py
@@ -15,12 +15,14 @@ def create_app(testing_config=None):
                pool_recycle=config.SQL_POOL_RECYCLE, pool_timeout=config.SQL_POOL_TIMEOUT,
                pool_pre_ping=config.SQL_POOL_PRE_PING, max_overflow=config.SQL_MAX_OVERFLOW)
 
-    from mishards.connections import ConnectionMgr
-    connect_mgr = ConnectionMgr()
+    from mishards.connections import ConnectionMgr, ConnectionTopology
+
+    readonly_topo = ConnectionTopology()
+    writable_topo = ConnectionTopology()
 
     from discovery.factory import DiscoveryFactory
     discover = DiscoveryFactory(config.DISCOVERY_PLUGIN_PATH).create(config.DISCOVERY_CLASS_NAME,
-                                                                     conn_mgr=connect_mgr)
+                                                                     readonly_topo=readonly_topo)
 
     from mishards.grpc_utils import GrpcSpanDecorator
     from tracer.factory import TracerFactory
@@ -30,12 +32,15 @@ def create_app(testing_config=None):
 
     from mishards.router.factory import RouterFactory
     router = RouterFactory(config.ROUTER_PLUGIN_PATH).create(config.ROUTER_CLASS_NAME,
-                                                             conn_mgr=connect_mgr)
+                                                             readonly_topo=readonly_topo,
+                                                             writable_topo=writable_topo)
 
-    grpc_server.init_app(conn_mgr=connect_mgr,
+    grpc_server.init_app(writable_topo=writable_topo,
+                         readonly_topo=readonly_topo,
                          tracer=tracer,
                          router=router,
-                         discover=discover)
+                         discover=discover,
+                         max_workers=settings.MAX_WORKERS)
 
     from mishards import exception_handlers